feat(ec2): Implement generating a pricing map
This is the next step towards implementing #200.
Updates logic in `pricing_map.go` to add an attribute for the total
price of the instance.
Largely copies the implementation in `eks.go` to generate the pricing
map by listing out OnDemand and Spot prices.
Refactors shared methods from `eks.go` to list on demand prices, spot
prices, and instances to `pricing_map.go` and `compute.go` respectively.
Create unit tests to ensure `ec2.Collect` has 100% coverage.
